Update UWP template
