--- ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/BlankApp.UWP/.template.config/template.json
@@ -1,20 +1,0 @@
-{
-    "author": "MonoGameTeam",
-    "classifications": [ "MonoGame" ],
-    "name": "MonoGame Windows Universal Application",
-    "groupIdentity": "MonoGame.BlankApp.UWP",
-    "identity": "MonoGame.BlankApp.UWP.CSharp",
-    "shortName": "mguwp",
-    "tags": {
-        "language": "C#",
-        "type": "project"
-    },
-    "primaryOutputs": [ 
-                        {
-                            "path": "BlankApp.UWP.csproj"
-                        },
-                      ], 
-    "sourceName": "BlankApp", 
-    "preferNameDirectory" : "true",
-    "description": "A MonoGame game project for Windows Universal using DirectX."
-}
