--- ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/BlankApp.UWP/App.xaml
@@ -1,8 +1,0 @@
-﻿<Application
-    x:Class="BlankApp.UWP.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:BlankApp.UWP"
-    RequestedTheme="Light">
-
-</Application>
