--- ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/BlankApp.UWP/Content/Content.mgcb
@@ -1,15 +1,0 @@
-
-#----------------------------- Global Properties ----------------------------#
-
-/outputDir:bin/$(Platform)
-/intermediateDir:obj/$(Platform)
-/platform:WindowsStoreApp
-/config:
-/profile:Reach
-/compress:False
-
-#-------------------------------- References --------------------------------#
-
-
-#---------------------------------- Content ---------------------------------#
-
