+++ ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/MonoGame.Application.UWP.CSharp/.template.config/dotnetcli.host.json
@@ -1,0 +1,8 @@
+{
+    "symbolInfo": {
+        "copyrightName": {
+            "longName": "framework",
+            "shortName": "f"
+        }
+    }
+}
