+++ ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/MonoGame.Application.UWP.CSharp/Assets/Square150x150Logo.scale-200.png
