+++ ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/MonoGame.Application.UWP.CSharp/Assets/Square44x44Logo.scale-200.png
