+++ ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/MonoGame.Application.UWP.CSharp/Assets/Square44x44Logo.targetsize-24_altform-unplated.png
